Release 1.43.x (#234)
* adding ignore option to project diff (#194)

#194 Experimental features, Option to ignore directories while generating diff

* [skip ci] Updated Version

* Add optional metadata version flag to the package:valid command

Users may specify a metadata API version to use for validation, by passing the
'apiversion' flag to the package:valid commmand. The corresponding metadata JSON
file needs to be in the resource folder, otherwise an error is thrown.

* Update metadata.json to API version 48

* [skip ci] Updated Version

* Issue #191 Support to new profiles fields in diff (#201)

#191 Support updates to profiles

* [skip ci] Updated Version

* Issue #196 Display Other Aura Components elements (#204)

#196 Display Lightning Event, Interface and App in the diff output

* [skip ci] Updated Version

* Update package.json for 1.38.X Milestone

* [skip ci] Updated Version

* Issue #202 Remove unnecessary parameters to diff (#203)

* [skip ci] Updated Version

* Update default PMD version to 6.21.0 (#207)

Check whether pmd_cache_directory already exists before creating directory. This
resolves the error that was preventing multiple PMD release versions to be
cached.

Add console log that informs the user which PMD release version is running.

* [skip ci] Updated Version

* Add support of CustomSettingAccess (#206)

Add CustomSettingAccess to profile merge and reconcile

* [skip ci] Updated Version

* Update diffUtil.ts (#210)

Issue #200
Fix Issue running git with renamed files

* [skip ci] Updated Version

* Issue #190 (#211)

Write a json output file
orgdiff.json in the wotking folder
diff.json in the output folder of the comand.


* [skip ci] Updated Version

* Update Version for upcoming release

* [skip ci] Updated Version

* Updated badges

* Issue/215 (#216)

Fix Issue #215

* [skip ci] Updated Version

* Update package.json

* [skip ci] Updated Version

* issue #217 (#218)

* issue #217

issue #217

exposting test result in csv or json format

* json output

* [skip ci] Updated Version

* Update version to 1.42.0

Prepping for 1.42.X milestone

* [skip ci] Updated Version

* Package Version Info with minor refactoring on other commands (#222)

* Package Version Info with minor refactoring on other commands

* show license details and usage

* update yarn.lock


* Issue/225 (#226)

Fix Layout name when building manifest


* Issue #219 (#227)

* Issue #219

Catch errors during diff handling and display them in the output

* Issue #219

Make the command fail when error occur


* medata info for entitlementProcess,milestoneType,fieldTranslation (#230)

fixes #229

* [skip ci] Updated Version

* Fix Apex Test Suite (#232)

* [skip ci] Updated Version

+++ src/commands/sfpowerkit/project/diff.ts
@@ -9,6 +9,8 @@ import DiffImpl from "../../../impl/project/diff/diffImpl";
 import * as path from "path";
 import { SFPowerkit } from "../../../sfpowerkit";
 import { fs } from "@salesforce/core";
+import rimraf = require("rimraf");
+import * as fsextra from "fs-extra";
 
 // Initialize Messages with the current plugin directory
 core.Messages.importMessagesDirectory(__dirname);
@@ -94,7 +96,9 @@ export default class Diff extends SfdxCommand {
     },
     display() {
       if (Array.isArray(this.data) && this.data.length) {
-        this.ux.table(this.data, this.tableColumnData);
+        this.ux.table(this.data.filter(element=>{
+          return element['action']!=='ERROR';
+        }), this.tableColumnData);
       }
     }
   };
@@ -121,7 +125,30 @@ export default class Diff extends SfdxCommand {
       this.flags.packagedirectories,
       this.flags.apiversion
     );
-    fs.writeJson(path.join(outputFolder, "diff.json"), diffOutput);
+    let errors = diffOutput.filter(element=>{
+      return element['action']==='ERROR';
+    });
+    if(errors && errors.length>0){
+      this.ux.log('ERRORS');
+      this.ux.table(errors, {
+        columns: [
+          { key: "path", label: "Path" },
+          { key: "message", label: "Error Message" }
+        ]
+      });
+      rimraf.sync(outputFolder);
+      if(this.flags.json){
+        //In case of error, the diff output is still printed in the output folder
+        if (fsextra.existsSync(outputFolder) == false) {
+          fsextra.mkdirSync(outputFolder);
+          fs.writeJson(path.join(outputFolder, "diff.json"), diffOutput);
+        }
+      }
+      throw new Error('Error parsing diff.');
+    }
+    else{
+      fs.writeJson(path.join(outputFolder, "diff.json"), diffOutput);
+    }
     return diffOutput;
   }
 }
